feat : authenticate identity with recognize face automation and validate automation identity

+++ src/main/java/com/east2west/service/HomestayService.java
@@ -11,6 +11,9 @@ import java.time.ZoneId;
 
 import com.east2west.models.DTO.*;
 import com.east2west.models.Entity.*;
+import com.east2west.models.enums.EHomestayStatus;
+import com.east2west.models.mapper.HomestayMapper;
+import com.east2west.models.mapper.StructureMapper;
 import com.east2west.repository.*;
 import com.east2west.util.DateUtil;
 import org.locationtech.jts.geom.Coordinate;
@@ -97,29 +100,30 @@ public class HomestayService {
     }
 
 
-    public Homestay createHomestay(HomestayDTO homestayDTO) {
+    public HomestayDTO createHomestay(HomestayDTO homestayDTO) {
 
+        homestayDTO.setIsApproved(false);
         Homestay homestay = new Homestay();
         populateHomestayFields(homestay, homestayDTO);
-
-
-
-            Homestay savedHomestay = homestayRepository.save(homestay);
+        if(homestayDTO.getPricePerNight() != null){
             LocalDate today = LocalDate.now();
             LocalDate endDate = today.plusDays(365);
-
+            List<HomestayAvailability> homestayAvailabilityList = new ArrayList<>();
             while (!today.isAfter(endDate)) {
                 HomestayAvailability availability = new HomestayAvailability();
-                availability.setHomestay(savedHomestay);
+                availability.setHomestay(homestayRepository.findById(homestayDTO.getHomestayid()).get());
                 availability.setDate(Timestamp.valueOf(today.atStartOfDay(ZoneId.systemDefault()).toLocalDateTime()));
                 availability.setStatus("Available");
                 availability.setPricepernight(homestayDTO.getPricePerNight());
 
-                homestayAvailabilityRepository.save(availability);
+                homestayAvailabilityList.add(availability);
                 today = today.plusDays(1);
 
             }
-        return savedHomestay;
+            homestay.setHomestayAvailabilityList(homestayAvailabilityList);
+        }
+            Homestay response = homestayRepository.save(homestay);
+        return HomestayMapper.INSTANCE.toDTO(response);
     }
 
     public List<HomestayDTO> search(String keyword,int userId) {
@@ -232,7 +236,7 @@ public class HomestayService {
         booking.setCheckin(bookingDTO.getCheckin());
         booking.setCheckout(bookingDTO.getCheckout());
         booking.setFeeamount(bookingDTO.getFeeamount());
-        if(availability.getHomestay().isInstant()){
+        if(availability.getHomestay().getInstant()){
             booking.setStatus("Booked");
 
         }else{
@@ -261,24 +265,63 @@ public class HomestayService {
 
 
     private void populateHomestayFields(Homestay homestay, HomestayDTO homestayDTO) {
+        if(homestayDTO.getHomestayid() != null){
+            homestay.setHomestayid(homestayDTO.getHomestayid());
+        }
         homestay.setUserid(homestayDTO.getUserId());
-        homestay.setLongitude(homestayDTO.getLongitude());
-        homestay.setLatitude(homestayDTO.getLatitude());
-        homestay.setTitle(homestayDTO.getTitle());
-        homestay.setAddress(homestayDTO.getAddress());
-        GeometryFactory geometryFactory = new GeometryFactory();
-        Point point = geometryFactory.createPoint(new Coordinate(homestayDTO.getLongitude(), homestayDTO.getLatitude()));
-        homestay.setGeom(point);
-        homestay.setPhotos(homestayDTO.getPhotos());
-        homestay.setDescription(homestayDTO.getDescription());
-        homestay.setExtrainfo(homestayDTO.getExtraInfo());
-        homestay.setCleaningfee(homestayDTO.getCleaningFee());
-        homestay.setIsapproved(homestayDTO.getIsApproved());
-        homestay.setMaxguest(homestayDTO.getMaxGuest());
+        if(homestayDTO.getLongitude() != null && homestayDTO.getLatitude() != null){
+            homestay.setLongitude(homestayDTO.getLongitude());
+            homestay.setLatitude(homestayDTO.getLatitude());
+            GeometryFactory geometryFactory = new GeometryFactory();
+            Point point = geometryFactory.createPoint(new Coordinate(homestayDTO.getLongitude(), homestayDTO.getLatitude()));
+            homestay.setGeom(point);
+        }
+        if(homestayDTO.getStatus() != null){
+            homestay.setStatus(homestayDTO.getStatus());
+        }
+        if(homestayDTO.getTitle() != null){
+            homestay.setTitle(homestayDTO.getTitle());
+        }
+        if(homestayDTO.getAddress() != null){
+            homestay.setAddress(homestayDTO.getAddress());
+        }
+        if(homestayDTO.getPhotos() != null){
+            homestay.setPhotos(homestayDTO.getPhotos());
+        }
+        if(homestayDTO.getDescription() != null){
+            homestay.setDescription(homestayDTO.getDescription());
+        }
+        if(homestayDTO.getExtraInfo() != null) {
+            homestay.setExtrainfo(homestayDTO.getExtraInfo());
+        }
+
+        if(homestayDTO.getCleaningFee() != null){
+            homestay.setCleaningfee(homestayDTO.getCleaningFee());
+        }
+
+        if(homestayDTO.getIsApproved() != null){
+            homestay.setIsapproved(false);
+        }
+
+
+        if(homestayDTO.getAmenityIds() != null && !homestayDTO.getAmenityIds().isEmpty()){
+            List<Amenities> amenities = amenitiesRepository.findByAmenitiesidIn(homestayDTO.getAmenityIds());
+            homestay.setAmenities(amenities);
+        }
+
+        if(homestayDTO.getMaxGuest() != null){
+            homestay.setMaxguest(homestayDTO.getMaxGuest());
+        }
         homestay.setType(homestayDTO.getType());
-        homestay.setBathroom(homestayDTO.getBathroom());
-        homestay.setBeds(homestayDTO.getBeds());
-        homestay.setInstant(homestayDTO.getInstant());
+        if(homestayDTO.getBathroom() != null){
+            homestay.setBathroom(homestayDTO.getBathroom());
+        }
+        if(homestayDTO.getBeds() != null){
+            homestay.setBeds(homestayDTO.getBeds());
+        }
+
+            homestay.setInstant(homestayDTO.getInstant());
+
         if(homestayDTO.getRoom() == null){
             homestay.setRoom(0);
         }else{
@@ -286,38 +329,37 @@ public class HomestayService {
         }
 
 
-
-        CityProvince cityProvince = new CityProvince();
-        cityProvince.setCityname(homestayDTO.getCityProvinceName());
-        cityProvince = cityProvinceRepository.save(cityProvince);
-
-
-        District district = new District();
-        district.setDistrictname(homestayDTO.getDistrictName());
-        district.setCityprovince(cityProvince);
-        district = districtRepository.save(district);
+        if(homestayDTO.getWardName() != null && homestayDTO.getCityProvinceName() != null && homestayDTO.getDistrictName() != null){
+            CityProvince cityProvince = new CityProvince();
+            cityProvince.setCityname(homestayDTO.getCityProvinceName());
+            cityProvince = cityProvinceRepository.save(cityProvince);
 
 
-        Ward ward = new Ward();
-        ward.setWardname(homestayDTO.getWardName());
-        ward.setDistrict(district);
-        ward = wardRepository.save(ward);
+            District district = new District();
+            district.setDistrictname(homestayDTO.getDistrictName());
+            district.setCityprovince(cityProvince);
+            district = districtRepository.save(district);
 
 
-        homestay.setWard(ward);
+            Ward ward = new Ward();
+            ward.setWardname(homestayDTO.getWardName());
+            ward.setDistrict(district);
+            ward = wardRepository.save(ward);
 
 
-        Structure structure = structureRepository.findById(homestayDTO.getStructureId())
-                .orElseThrow(() -> new ResourceNotFoundException(
-                        "structure not found with id " + homestayDTO.getStructureId()));
-        homestay.setStructure(structure);
+            homestay.setWard(ward);
+        }
 
+        if(homestayDTO.getStructureId() != null){
+            Structure structure = structureRepository.findById(homestayDTO.getStructureId())
+                    .orElseThrow(() -> new ResourceNotFoundException(
+                            "structure not found with id " + homestayDTO.getStructureId()));
+            homestay.setStructure(structure);
 
-        if (homestayDTO.getPerkIds() != null && !homestayDTO.getPerkIds().isEmpty()) {
-            List<Amenities> amenities = amenitiesRepository.findAllById(homestayDTO.getPerkIds());
+        }
+        if (homestayDTO.getAmenityIds() != null && !homestayDTO.getAmenityIds().isEmpty()) {
+            List<Amenities> amenities = amenitiesRepository.findAllById(homestayDTO.getAmenityIds());
             homestay.setAmenities(amenities);
-        } else {
-            homestay.setAmenities(new ArrayList<>());
         }
     }
 
@@ -378,11 +420,25 @@ public HomestayDTO getHomestayHomestayAvailabilityById(int id){
     }
 
 
-    public void  deleteHomestay(int id) {
-       List<HomestayAvailability>  ava =  homestayAvailabilityRepository.findByHomestay_Homestayid(id);
-        homestayAvailabilityRepository.deleteAll(ava);
-        homestayRepository.deleteById(id);
+
+    public String deleteHomestay(int id){
+
+        Optional<Homestay> homestay = homestayRepository.findById(id);
+        if(homestay.isPresent()){
+            List<HomestayAvailability>  ava =  homestayAvailabilityRepository.findByHomestay_Homestayid(id);
+            homestayAvailabilityRepository.deleteAll(ava);
+            homestayRepository.deleteById(id);
+            return "Deleted " + homestay.get().getTitle() + " homestay successfully";
+        }else{
+            return "Not found homestay";
+        }
     }
+
+    public List<HomestayDTO> getAllHomestayStatus(int userId , EHomestayStatus status) {
+        List<Homestay> homestays = homestayRepository.findByUseridAndStatus(userId, status);
+        return homestays.stream().map(HomestayMapper.INSTANCE::toDTO).toList();
+    }
+
 
 
     public List<HomestayDTO> getAllApproved() {
@@ -419,7 +475,7 @@ public HomestayDTO getHomestayHomestayAvailabilityById(int id){
     public List<HomestayDTO> getAllByIdUser(int id) {
         List<Homestay> homestays = homestayRepository.findByUserid(id);
         return homestays.stream()
-                .map(this::convertToDTO)
+                .map(HomestayMapper.INSTANCE::toDTO)
                 .collect(Collectors.toList());
     }
 
@@ -442,8 +498,8 @@ public HomestayDTO getHomestayHomestayAvailabilityById(int id){
         dto.setType(homestay.getType());
         dto.setBeds(homestay.getBeds());
         dto.setBathroom(homestay.getBathroom());
-        dto.setInstant(homestay.isInstant());
-        dto.setPerkIds(homestay.getAmenities().stream()
+        dto.setInstant(homestay.getInstant());
+        dto.setAmenityIds(homestay.getAmenities().stream()
                 .map(Amenities::getAmenitiesid)
                 .collect(Collectors.toList()));
 
@@ -464,14 +520,14 @@ public HomestayDTO getHomestayHomestayAvailabilityById(int id){
         dto.setDistrictName(homestay.getWard().getDistrict().getDistrictname());
         dto.setCityProvinceName(homestay.getWard().getDistrict().getCityprovince().getCityname());
         // TODO: Check data type here
-        if (homestay.getGeom() != null) {
-            WKBWriter wkbWriter = new WKBWriter();
-            byte[] wkb = wkbWriter.write(homestay.getGeom());
-            String wkbHex = bytesToHex(wkb);
-            dto.setGeom(wkbHex);
-        }
+//        if (homestay.getGeom() != null) {
+//            WKBWriter wkbWriter = new WKBWriter();
+//            byte[] wkb = wkbWriter.write(homestay.getGeom());
+//            String wkbHex = bytesToHex(wkb);
+//            dto.setGeom(wkbHex);
+//        }
 
-        //        if (homestay.getGeom() != null) {
+//                if (homestay.getGeom() != null) {
 //            try {
 //                WKTReader wktReader = new WKTReader();
 //                Geometry geometry = wktReader.read(homestay.getGeom());
@@ -501,10 +557,15 @@ public HomestayDTO getHomestayHomestayAvailabilityById(int id){
         return convertToDTO(homestay);
     }
 
-    public Homestay getHomestayById(int id){
-        return homestayRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Homestay not found with id " + id));
+
+
+    public Optional<HomestayDTO> getHomestayById(int id){
+        Optional<Homestay> homestay = homestayRepository.findById(id);
+        if(homestay.isEmpty()) return Optional.empty();
+        HomestayDTO response = HomestayMapper.INSTANCE.toDTO(homestay.get());
+        return Optional.ofNullable(response);
     }
+
 
     public void updateBasePrice(int homestayId, BigDecimal newPrice) {
         homestayAvailabilityRepository.updateBasePrice(homestayId, newPrice);
